# backend/src/swaggerSpec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Local Resource Sharing API",
    "version": "1.0.0",
    "description": "API documentation for the Local Resource Sharing platform backend."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/marketplace/categories": {
      "get": {
        "summary": "Get all resource categories with availability counts",
        "tags": [
          "Marketplace"
        ],
        "responses": {
          "200": {
            "description": "List of categories with counts"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/marketplace/{category}": {
      "get": {
        "summary": "Get helpers and organizations offering a specific resource category",
        "tags": [
          "Marketplace"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Category of resource (e.g., blood, food, transport, shelter)"
          },
          {
            "in": "query",
            "name": "lat",
            "schema": {
              "type": "number"
            },
            "description": "Latitude for geospatial sorting"
          },
          {
            "in": "query",
            "name": "lng",
            "schema": {
              "type": "number"
            },
            "description": "Longitude for geospatial sorting"
          },
          {
            "in": "query",
            "name": "radiusInKm",
            "schema": {
              "type": "number"
            },
            "description": "Max distance in km (default 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching users"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}